feat: 설정을 위한 provider들을 storybook decorator에 추가 #25
